requirements.txt
asgiref==3.2.10
astroid==2.4.2
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.2.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.43.0
factory-boy==3.1.0
Faker==4.4.0
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.5.4
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.0
prompt-toolkit==3.0.7
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
traitlets==5.0.4
urllib3==1.25.11
wcwidth==0.2.5
wrapt==1.12.1
